add aiter tests and move json_dump header
