+++ example/ck_tile/01_fmha/fmha_fwd.cpp
@@ -7,7 +7,7 @@
 #include "mask.hpp"
 #include "rotary.hpp"
 #include "utils.hpp"
-#include "json_dump.hpp"
+#include "ck_tile/utility/json_dump.hpp"
 
 #include <array>
 #include <cstring>
