+++ example/ck_tile/10_rmsnorm2d/rmsnorm2d_fwd.cpp
@@ -1,7 +1,7 @@
 #include "ck_tile/host.hpp"
 #include "rmsnorm2d_fwd.hpp"
 #include <cstring>
-#include "json_dump.hpp"
+#include "ck_tile/utility/json_dump.hpp"
 
 // different threshold for different dtype
 template <typename DataType>
