+++ example/ck_tile/14_moe_smoothquant/moe_smoothquant.cpp
@@ -1,6 +1,6 @@
 #include "ck_tile/host.hpp"
 #include "moe_smoothquant.hpp"
-#include "json_dump.hpp"
+#include "ck_tile/utility/json_dump.hpp"
 #include <cstring>
 #include <set>
