+++ example/ck_tile/15_fused_moe/main.cpp
@@ -5,7 +5,7 @@
 #include <set>
 
 #include "ck_tile/host.hpp"
-#include "json_dump.hpp"
+#include "ck_tile/utility/json_dump.hpp"
 #include "fused_moe.hpp"
 
 // different threshold for different dtype
