+++ example/ck_tile/19_gemm_multi_d/run_gemm_multi_d_fp16_example.inc
@@ -3,7 +3,7 @@
 
 #pragma once
 #include <cstddef>
-#include "json_dump.hpp"
+#include "ck_tile/utility/json_dump.hpp"
 
 template <typename ADataType,
           typename BDataType,
